move django to lib/charms/layer
